+++ Classes/Utils/UIImage+RGBAsAtXandCount.m
@@ -33,8 +33,12 @@
     CGContextRelease(context);
 	
     // Now your rawData contains the image data in the RGBA8888 pixel format.
-    int byteIndex = (bytesPerRow * yy) + xx * bytesPerPixel;
-    for (int ii = 0 ; ii < count ; ++ii)
+// begin cbleu fix
+//	int byteIndex = (bytesPerRow * yy) + xx * bytesPerPixel;
+	unsigned long byteIndex = (bytesPerRow * yy) + xx * bytesPerPixel;
+// end cbleu fix
+
+	for (int ii = 0 ; ii < count ; ++ii)
     {
         CGFloat red   = (rawData[byteIndex]     * 1.0) / 255.0;
         CGFloat green = (rawData[byteIndex + 1] * 1.0) / 255.0;
